change class-components to function-components
