+++ src/App.jsx
@@ -1,15 +1,16 @@
 import React, {Component}from 'react';
 import Navbar from './components/Navbar';
-import Input from './components/Input';
-import Output from './components/Output';
+import InputDish from './components/InputDish';
+import OutputDish from './components/OutputDish';
 import "bootstrap/dist/css/bootstrap.min.css";
+
 
 function App() {
   return (
     <div className="App">
       <Navbar />
-      <Input />
-      <Output />
+      <InputDish />
+      <OutputDish />
     </div>
   );
 }
