--- src/components/Input.jsx
@@ -1,19 +1,0 @@
-import React, {Component} from "react";
-
-
-class Input extends Component {
-
-    state = {textareaValue: ''};
-  
-    handleTextareaChange = (e) => {
-      this.setState({textareaValue: e.target.value });
-    }
-  
-    render() {
-      return (
-          <textarea className="Eingabefeld" name="Eingabefeld" placeholder="Bitte Zutaten eingeben" value={this.state.textareaValue} onChange={this.handleTextareaChange}/>
-      );
-    }
-  }
- 
-export default Input;
