+++ src/components/Navbar.jsx
@@ -1,13 +1,22 @@
-import React, {Component} from "react";
+/*import React, {Component} from "react";*/
 
 
-class Navbar extends React.Component {
+function NavBar (props) {
+  return (
+    <nav>Navigationbar</nav>
+  );
+}
+
+export default NavBar;
+
+
+/* class Navbar extends React.Component {
 
     render() {
       return (
-          <nav>Navigationbar</nav>
+          
       );
     }
   }
  
-export default Navbar;
+export default Navbar; */
